refactor: 의미있는 서비스명으로 변경 - 각 Entitiy에 @SQLRestriction 추가 - BaseEntity에 delete()추가
