+++ tuk-api/src/main/kotlin/nexters/tuk/application/gathering/GatheringService.kt
@@ -1,77 +1,40 @@
 package nexters.tuk.application.gathering
 
 import nexters.tuk.application.gathering.dto.request.GatheringCommand
-import nexters.tuk.application.gathering.dto.request.GatheringQuery
-import nexters.tuk.application.gathering.dto.response.GatheringFacadeResponse
-import nexters.tuk.application.invitation.InvitationService
-import nexters.tuk.application.member.MemberService
+import nexters.tuk.application.gathering.dto.response.GatheringResponse
+import nexters.tuk.domain.gathering.Gathering
+import nexters.tuk.domain.gathering.GatheringRepository
+import nexters.tuk.domain.gathering.findByIdOrThrow
 import org.springframework.stereotype.Service
-import org.springframework.transaction.annotation.Transactional
+import java.time.LocalDate
+import java.time.temporal.ChronoUnit
 
 @Service
 class GatheringService(
-    private val gatheringDomainService: GatheringDomainService,
-    private val gatheringMemberService: GatheringMemberService,
-    private val invitationService: InvitationService,
-    private val memberService: MemberService,
+    private val gatheringRepository: GatheringRepository,
 ) {
-    @Transactional
-    fun generateGathering(command: GatheringCommand.Generate): GatheringFacadeResponse.Generate {
-        val gatheringId = gatheringDomainService.generateGathering(command)
+    fun generateGathering(command: GatheringCommand.Generate): Long {
+        val gathering = Gathering.generate(command).also { gatheringRepository.save(it) }
 
-        gatheringMemberService.registerMember(
-            gatheringId = gatheringId,
-            memberId = command.memberId,
-        )
+        return gathering.id
+    }
 
-        // TODO 알림 등록하기
-        return GatheringFacadeResponse.Generate(
-            gatheringId,
+    fun getGatheringDetail(gatheringId: Long): GatheringResponse.GatheringDetail {
+        val gathering = gatheringRepository.findByIdOrThrow(gatheringId)
+
+        return GatheringResponse.GatheringDetail(
+            gathering.id,
+            gathering.name,
+            gathering.firstGatheringDate.daysAgo(),
+            gathering.lastGatheringDate.monthsAgo(),
         )
     }
 
-    @Transactional(readOnly = true)
-    fun getMemberGatherings(query: GatheringQuery.MemberGathering): GatheringFacadeResponse.GatheringOverviews {
-        val gatherings = gatheringMemberService.getMemberGatherings(query.memberId)
-        val gatheringOverviews = gatherings.map {
-            GatheringFacadeResponse.GatheringOverviews.GatheringOverview(
-                it.id, it.name, it.monthsSinceLastGathering
-            )
-        }
-
-        return GatheringFacadeResponse.GatheringOverviews(gatherings.size, gatheringOverviews)
+    private fun LocalDate.daysAgo(): Int {
+        return until(LocalDate.now(), ChronoUnit.DAYS).toInt()
     }
 
-    @Transactional(readOnly = true)
-    fun getGatheringDetail(query: GatheringQuery.GatheringDetail): GatheringFacadeResponse.GatheringDetail {
-        gatheringMemberService.verifyGatheringAccess(query.gatheringId, query.memberId)
-
-        val invitationStat = invitationService.getGatheringInvitationStat(query.gatheringId, query.memberId)
-
-        val gatheringMemberIds = gatheringMemberService.getGatheringMemberIds(query.gatheringId)
-        val members = memberService.getMemberOverview(gatheringMemberIds).map {
-                GatheringFacadeResponse.GatheringDetail.MemberOverview(it.memberId, it.memberName)
-            }
-
-        val gatheringDetail = gatheringDomainService.getGatheringDetail(query.gatheringId)
-
-        return GatheringFacadeResponse.GatheringDetail(
-            gatheringId = gatheringDetail.id,
-            gatheringName = gatheringDetail.name,
-            daysSinceFirstGathering = gatheringDetail.daysSinceFirstGathering,
-            monthsSinceLastGathering = gatheringDetail.monthsSinceLastGathering,
-            sentInvitationCount = invitationStat.sentCount,
-            receivedInvitationCount = invitationStat.receivedCount,
-            members = members
-        )
-    }
-
-    @Transactional
-    fun joinGathering(command: GatheringCommand.JoinGathering): GatheringFacadeResponse.JoinGathering {
-        gatheringMemberService.registerMember(
-            gatheringId = command.gatheringId, memberId = command.memberId
-        )
-
-        return GatheringFacadeResponse.JoinGathering(command.gatheringId)
+    private fun LocalDate.monthsAgo(): Int {
+        return until(LocalDate.now(), ChronoUnit.MONTHS).toInt()
     }
 }
