+++ tuk-api/src/main/kotlin/nexters/tuk/domain/invitation/Invitation.kt
@@ -3,8 +3,11 @@ package nexters.tuk.domain.invitation
 import jakarta.persistence.*
 import nexters.tuk.domain.BaseEntity
 import nexters.tuk.domain.gathering.Gathering
+import org.hibernate.annotations.SQLRestriction
 
 // FIXME: 모임 생성을 위한 임시 초대장 엔티티
+@SQLRestriction("deleted_at is NULL")
+@Table(name = "invitation")
 @Entity
 class Invitation(
     @Column(name = "member_id", nullable = false)
